load only necessary proto files when testing a filter with proto transformers
