+++ contracts/external/dao-rbam/src/error.rs
@@ -25,8 +25,8 @@ pub enum ContractError {
     #[error("JSON serialization error: {err}")]
     JsonSerialization { err: String },
 
-    #[error("Invalid filter: {err}")]
-    FilterInvalid { err: String },
+    #[error("Filter error: {err}")]
+    FilterError { err: String },
 
     #[error("Message not allowed by filter: {err}")]
     MsgNotAllowedByFilter { err: String },
@@ -84,14 +84,27 @@ pub enum ContractError {
         file_descriptor_set_index: usize,
     },
 
-    #[error("Message descriptor {message_descriptor_index} missing name in file {file_name}")]
+    #[error(
+        "File descriptor {file_descriptor_index} (name: {file_descriptor_name}) missing package in set {file_descriptor_set_index}"
+    )]
+    FileDescriptorMissingPackage {
+        file_descriptor_index: usize,
+        file_descriptor_name: String,
+        file_descriptor_set_index: usize,
+    },
+
+    #[error("Message descriptor {message_descriptor_index} missing name in file {file_name} (package: {file_package})")]
     MessageDescriptorMissingName {
         message_descriptor_index: usize,
         file_name: String,
+        file_package: String,
     },
 
     #[error("Message limit reached before all files were unregistered ({unregistered}/{total}). Increase the limit or unregister fewer files.")]
     ProtobufMessageLimitReached { unregistered: usize, total: usize },
+
+    #[error("Protobuf message `{message}` not registered")]
+    ProtobufMessageNotFound { message: String },
 }
 
 impl From<ContractError> for String {
